+++ sktransf/log.py
@@ -7,7 +7,10 @@ import pandas as pd
 from sklearn.base import BaseEstimator, TransformerMixin
 
 
-class LogTransformer(BaseEstimator, TransformerMixin):
+pd.set_option("future.no_silent_downcasting", True)
+
+
+class LogColumnTransformer(BaseEstimator, TransformerMixin):
     """Logarithm transformer
 
     Opt args: