+++ tests/unit/test_unique.py
@@ -2,6 +2,9 @@
 Unique Bool transformer
 """
 
+import pytest
+
+import numpy as np
 import pandas as pd
 
 from sktransf import DropUniqueColumnTransformer
@@ -21,13 +24,13 @@ class TestDropUniqueColumnTransformer:
         """Test the init method"""
 
         # create the transformer
-        transformer = DropUniqueColumnTransformer()
+        transformer = DropUniqueColumnTransformer(force_df_out=True)
 
         assert transformer.unique_cols is None
 
     def test_fit(self, X_unique: pd.DataFrame):
         # create the transformer
-        transformer = DropUniqueColumnTransformer()
+        transformer = DropUniqueColumnTransformer(force_df_out=True)
 
         # fit the transformer
         transformer.fit(X_unique)
@@ -36,7 +39,7 @@ class TestDropUniqueColumnTransformer:
 
     def test_transform(self, X_unique: pd.DataFrame):
         # create the transformer
-        transformer = DropUniqueColumnTransformer()
+        transformer = DropUniqueColumnTransformer(force_df_out=True)
 
         # fit the transformer
         transformer.fit(X_unique)
@@ -47,3 +50,28 @@ class TestDropUniqueColumnTransformer:
         # assert
         assert isinstance(X_unique_, pd.DataFrame)
         assert "unique_col" not in X_unique_.columns
+
+    @pytest.mark.parametrize(
+        "force_df_out,_type",
+        [
+            (True, pd.DataFrame),
+            (False, np.ndarray),
+        ],
+    )
+    def test_force_df_out(
+        self,
+        X: pd.DataFrame,
+        force_df_out: bool,
+        _type: None,
+    ):
+        """Test the force_df_out attribute"""
+
+        # create the transformer
+        transformer = DropUniqueColumnTransformer(force_df_out=force_df_out)
+
+        # fit the transformer
+        transformer.fit(X)
+        X_ = transformer.transform(X)
+
+        # type  check
+        assert isinstance(X_, _type)