+++ docs/docs/CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 
-## -- Not released -- [0.2.1] - 2024-02-08
+## [0.2.1] - 2024-02-09
 
 ### Added
 - Full support for official documentation
 - Add auto deploy documentation with GitHub pages
+- Add notebooks to the documentation
 
 ### Updated
 - Moove assets and utils to docs folder
@@ -17,6 +18,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Update documentation to include usage examples
 - Moove Changelog and Contributing to the docs folder
 - Clean code files, remove unused imports, useless comments, and unused variables
+- LogTransformer is now LogColumnTransformer
+- BoolColumnTransformer and DropUniqueColumnTransformer now accept force_df_out attribute
+- Rename modules logger to log
 
 ## [0.2.0] - 2024-02-08